% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DatavarMtcars}
\alias{DatavarMtcars}
\title{Datavar for the mtcars dataset in \code{datasets}}
\format{
A data.frame with 11 rows and 12 variables:
\describe{
\item{var}{The name of each variable. All variables you want to describe should be in the datavar.}
\item{type}{Type of variable : quanti for a quantitative variable; quali for a categorial variable with more than 2 items; and binary for a binary variable.}
\item{conf_inter}{For categorical and binary variables, the type of confidence interval for percentage you want to display ("none", "exact", "normal", "jeffreys").}
\item{conf_level}{For categorical and binary variables, the width of confidence interval if conf_inter isn't "none".}
\item{prec}{Number of decimals (for quantitative variables it is for the display of statistics, and for qualitative/binary variables it is for the percentages).}
\item{nomcateg}{For binary variables, the name of the class in the data to be displayed.}
\item{label}{For binary variables, the name displayed in the description table.}
\item{nomvariable}{For categorical and quantitative variables, the name of the variable displayed in the description table.}
\item{ordonnee}{For categorical variables, TRUE if you want to let the classes in order and FALSE if you want them by descending count.}
\item{mode}{For quantitative variables, string indicating which statistics to be displayed: moy for mean; med for median; sd for standard error; iq for interquartile interval; and rg for range.}
\item{test}{String indicating the statistical test to perform: none for none; student for Student's test; wilcoxon for Wilcoxon's signed rank test; ztest for Z-test; chisq for Chi-2 test; and fisher for Fisher's test.}
\item{chif_pval}{Number of decimals for Pvalues.}
}
}
\usage{
DatavarMtcars
}
\description{
A data.frame used by function \code{Description} to have all the arguments.
}
\keyword{datasets}
