% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDatavar.R
\name{DefaultOptions}
\alias{DefaultOptions}
\title{DefaultOptions}
\usage{
DefaultOptions(
  LimiteDetection = 10L,
  ConfInter = "none",
  ConfLevel = 0.95,
  Prec = NULL,
  Ordonnee = TRUE,
  Mode = "mediqrg",
  Test = "none",
  ChifPval = 2
)
}
\arguments{
\item{LimiteDetection}{Limit detection for the number of different items in a numeric variable to be considered quantitative.}

\item{ConfInter}{String for categorical variables indicating which type of confidence interval should be computed for percentage.}

\item{ConfLevel}{Numeric between 0 and 1 for categorical variables indicating the level for confidence intervals for percentages.}

\item{Prec}{Number of decimals for quantitative variables.}

\item{Ordonnee}{TRUE/FALSE for argument ordonnee of function \code{tab_quali}.}

\item{Mode}{mode argument of function \code{tab_quanti}.}

\item{Test}{Statistical test to be performed.}

\item{ChifPval}{Number of decimals of Pvalues.}
}
\value{
A list of default options.
}
\description{
Default options for the function \code{CreateDatavar}.
}
\examples{
print(DefaultOptions(mtcars))
}
\seealso{
\code{\link[=CreateDatavar]{CreateDatavar()}}
}
