% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{descr}
\alias{descr}
\title{descr}
\usage{
descr(
  data,
  y = NULL,
  datavar,
  listevar = datavar[[1]],
  nomcol = NULL,
  langue = "fr",
  comparer = TRUE
)
}
\arguments{
\item{data}{The dataset.}

\item{y}{The crossing variable. NULL if you only want univariate description.}

\item{datavar}{The datavar.}

\item{listevar}{A vector of the name of the columns you want to describe. By default, it is the whole set of variables present in the datavar.}

\item{nomcol}{The vector of the names you want to give to the columns of the output. NULL for automatic naming.}

\item{langue}{"fr" for French and "eng" for English.}

\item{comparer}{TRUE won't alterate the datavar, but FALSE will ensure that no test will be performed even if supplied in the datavar. In case of crossed description.}
}
\description{
Description of a data.frame
}
\details{
A function using the 3 other functions (\code{tab_quanti}, \code{tab_quali} and \code{tab_binaire}) to describe the variables of a data.frame.

You can also use it to cross it with a categorial variable and perform a comparison test (chisq or fisher test). For now no tests are performed with crossing with a variable that has more than 2 classes.

The function is now updated as function \code{Description} and one should use the later instead.
}
\examples{
# The description of the whole mtcars dataset
descr(mtcars, datavar = datavarr, langue = "eng")
# Only some variables
descr(mtcars, datavar = datavarr, listevar = c("mpg", "am", "cyl"), langue = "eng")
# Crossed (messages for Wilcoxon's test because of ex aequos)
descr(mtcars, datavar = datavarr, y = am, langue = "eng", comparer = FALSE)
descr(mtcars, datavar = datavarr, y = am, langue = "eng")
}
\seealso{
\code{\link{tab_quali}}, \code{\link{tab_quanti}}, \code{\link{tab_binaire}}, \code{\link{Description}}
}
