% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{descr_var}
\alias{descr_var}
\title{descr_var}
\usage{
descr_var(data, ligne_datavar, y, nomcol, langue)
}
\arguments{
\item{data}{Data.frame of the data.}

\item{ligne_datavar}{The line in datavar of the variable to describe.}

\item{y}{The variable used to cross.}

\item{nomcol}{Vector of the names of the columns.}

\item{langue}{Language.}
}
\description{
descr_var
}
