% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description.R
\name{Description}
\alias{Description}
\alias{NoMessDescription}
\alias{SilentDescription}
\title{Description}
\usage{
Description(
  .Data,
  y = NULL,
  .Datavar,
  .Listevar = .Datavar[[1]],
  PMissing = NULL,
  Poids = NULL,
  SMD = FALSE,
  NomCol = NULL,
  Langue = "eng",
  Grapher = FALSE,
  Simplif = TRUE,
  Comparer = FALSE
)

NoMessDescription(
  .Data,
  y = NULL,
  .Datavar,
  .Listevar = .Datavar[[1]],
  PMissing = NULL,
  Poids = NULL,
  SMD = FALSE,
  NomCol = NULL,
  Langue = "eng",
  Grapher = FALSE,
  Simplif = TRUE,
  Comparer = TRUE
)

SilentDescription(
  .Data,
  y = NULL,
  .Datavar,
  .Listevar = .Datavar[[1]],
  PMissing = NULL,
  Poids = NULL,
  SMD = FALSE,
  NomCol = NULL,
  Langue = "eng",
  Grapher = FALSE,
  Simplif = TRUE,
  Comparer = FALSE
)
}
\arguments{
\item{.Data}{The dataset to describe.}

\item{y}{The crossing variable. NULL if you only want univariate description.}

\item{.Datavar}{The datavar.}

\item{.Listevar}{A vector of the name of the columns you want to describe. By default, it is the whole set of variables present in the datavar.}

\item{PMissing}{Number of decimals of percentage from whole dataset. NULL (default value) if no percent desired.}

\item{Poids}{Name of the column of .Data in which the weights are stored. Let NULL for unweighted analysis.}

\item{SMD}{Boolean to indicate if you want standardized mean differences. Of note, for weighted analysis, only SMD are available and no test.}

\item{NomCol}{The vector of the names you want to give to the columns of the output. NULL for automatic naming.}

\item{Langue}{"fr" for French and "eng" for English.}

\item{Grapher}{Boolean if you want to graph the distribution in univariate case.}

\item{Simplif}{Not usefull, if TRUE will delete unused column 'Pvalue'.}

\item{Comparer}{TRUE won't alterate the datavar, but FALSE will ensure that no test will be performed even if supplied in the datavar. In case of crossed description. Of note, if no test is supplied in the datavar, even with Comparer = TRUE, there will be no comparison made.}
}
\description{
Description of a data.frame
}
\details{
A function using the other individual description functions (\code{TabQuanti}, \code{TabQuali} and \code{TabBinaire}) to describe the variables of a data.frame.

You can also use it to cross it with a categorial variable and perform a comparison test. For now no tests are performed with crossing with a variable that has more than 2 classes.

NoMessDescription removes the information messages and SilentDescription removes also warnings.
}
\examples{
# Create the datavar object
DatavarVoitures <- CreateDatavar(mtcars)
# The description of the whole mtcars dataset
Description(mtcars, .Datavar = DatavarVoitures, Langue = "eng")
# Only some variables
Description(mtcars, .Datavar = DatavarVoitures, .Listevar = c("mpg", "am", "cyl"), Langue = "eng")
# Crossed (messages for Wilcoxon's test because of ex aequos)
Description(mtcars, .Datavar = DatavarVoitures, y = am, Langue = "eng", Comparer = FALSE)
Description(mtcars, .Datavar = DatavarVoitures, y = am, Langue = "eng", Comparer = TRUE)
}
\seealso{
\code{\link[=TabQuali]{TabQuali()}}, \code{\link[=TabQuanti]{TabQuanti()}}, \code{\link[=TabBinaire]{TabBinaire()}}
}
