% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datavar.R
\name{detection_type}
\alias{detection_type}
\title{detection_type}
\usage{
detection_type(data, variable, limite_detection = 10L)
}
\arguments{
\item{data}{Dataset.}

\item{variable}{Name of the variable.}

\item{limite_detection}{Limit of number of items in the variable to be considered quantitative or qualitative.}
}
\description{
detection_type
}
