% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDatavar.R
\name{DetectionType}
\alias{DetectionType}
\title{DetectionType}
\usage{
DetectionType(.Data, .Var, LimiteDetection = 10L)
}
\arguments{
\item{.Data}{Dataset.}

\item{.Var}{Name of the variable.}

\item{LimiteDetection}{Limit of number of items in the variable to be considered quantitative or qualitative.}
}
\description{
DetectionType
}
