% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{ExtraireSMDs}
\alias{ExtraireSMDs}
\title{Extract SMDs}
\usage{
ExtraireSMDs(TabDescription)
}
\arguments{
\item{TabDescription}{tab_description object created by function \code{Description()}.}
}
\value{
Data.frame of SMDs for all variable in TabDescription object.
}
\description{
Extract SMDs
}
\examples{
TabDesc <- Description(mtcars, y = am, .Datavar = DatavarMtcars, SMD = TRUE)
ExtraireSMDs(TabDesc)
}
