% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{FlexTabDescr}
\alias{FlexTabDescr}
\title{Display results of description}
\usage{
FlexTabDescr(TabDescription, Widths = NULL, BaseSize = 10)
}
\arguments{
\item{TabDescription}{tab_description object.}

\item{Widths}{The widths of the displayed columns.}

\item{BaseSize}{Size of police.}
}
\value{
A flextable object.
}
\description{
Create a flextable object to display the result of a tab_description object.
}
\examples{
# Create the datavar object
DatavarVoitures <- CreateDatavar(mtcars)
# The description of the whole mtcars dataset
TabDescr <- Description(mtcars, .Datavar = DatavarVoitures, Langue = "eng", Grapher = TRUE)
# Display as flextable
FlexTabDescr(TabDescr)
}
