% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphes.R
\name{GGBar}
\alias{GGBar}
\title{Bar chart for categorical variables}
\usage{
GGBar(VarBinaire, VarCroise = NULL, Poids, Prec)
}
\arguments{
\item{VarBinaire}{Variable to plot.}

\item{VarCroise}{Crossing variable.}

\item{Poids}{Vector of weights}

\item{Prec}{Number of digits for percentages (not displayed now by choice).}
}
\value{
Single bar plot.
}
\description{
Bar chart for categorical variables
}
