% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphes.R
\name{GGHist}
\alias{GGHist}
\title{Histogramm for continuous variables}
\usage{
GGHist(VarQuanti, Bins = 10)
}
\arguments{
\item{VarQuanti}{Continuous variables.}

\item{Bins}{Number of bins}
}
\value{
Histogramm plot.
}
\description{
Histogramm for continuous variables
}
