% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{LargeBorder}
\alias{LargeBorder}
\title{Borders and formatting for flextable}
\format{
An object of class \code{fp_border} of length 3.
}
\usage{
LargeBorder
}
\description{
Borders and formatting for flextable
}
\keyword{datasets}
