% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{options_datavar}
\alias{options_datavar}
\title{options_datavar}
\usage{
options_datavar(
  limite_detection = 10L,
  coefbin = FALSE,
  prec = 2L,
  ordonnee = TRUE,
  mode = "mediqrg",
  test = "none",
  chif_pval = 2
)
}
\arguments{
\item{limite_detection}{Limit detection for the number of different items in a numeric variable to be considered quantitative.}

\item{coefbin}{Boolean for the display of the confidence interval of percentages.}

\item{prec}{Number of decimals for quantitative variables.}

\item{ordonnee}{TRUE/FALSE for argument ordonnee of function \code{tab_quali}.}

\item{mode}{mode argument of function \code{tab_quanti}.}

\item{test}{Statistical test to be performed.}

\item{chif_pval}{Number of decimals of Pvalues.}
}
\value{
A list of default options.
}
\description{
Default options for the function \code{create_datavar}
}
