% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{print.tab_description}
\alias{print.tab_description}
\title{Print description table in console}
\usage{
\method{print}{tab_description}(x, ...)
}
\arguments{
\item{x}{The table of description}
}
\description{
Print description table in console
}
\examples{
TabBinaire(mtcars, am, Grapher = TRUE)
}
