% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{ShapiroTest}
\alias{ShapiroTest}
\title{Wrapper function because of the error of shapiro.test with more than 5000 observations.}
\usage{
ShapiroTest(x, y = NULL, NameX)
}
\arguments{
\item{x}{Variable to test normality.}

\item{y}{Independant variable if there is one, else let NULL}

\item{NameX}{Name of variable.}
}
\value{
Pvalue
}
\description{
Wrapper function because of the error of shapiro.test with more than 5000 observations.
}
