% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{SmdMoy}
\alias{SmdMoy}
\title{SMD means}
\usage{
SmdMoy(Vec1, Poids1, Vec2, Poids2)
}
\arguments{
\item{Vec1, Poids1}{Values and weights of quantitative variable in group 1}

\item{Vec2, Poids2}{Values and weights of quantitative variable in group 2}
}
\value{
SMD for quantitative variable between 2 groups
}
\description{
SMD means
}
