% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{SmdProp}
\alias{SmdProp}
\title{SMD proportions}
\usage{
SmdProp(Vec1, Poids1, Vec2, Poids2)
}
\arguments{
\item{Vec1, Poids1}{Values and weights of categorical variable in group 1}

\item{Vec2, Poids2}{Values and weights of categorical variable in group 2}
}
\value{
SMD for categorical variable between the 2 groups
}
\description{
SMD proportions
}
