% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{tab_binaire}
\alias{tab_binaire}
\title{tab_binaire}
\usage{
tab_binaire(
  data,
  x,
  y = NULL,
  Prec = 0,
  coefbin = FALSE,
  chif_pval = 2,
  test = "none",
  Langue = "eng",
  nomcol = NULL,
  nomcateg = NULL,
  label = NULL,
  simplif = TRUE
)
}
\arguments{
\item{data}{The dataset that contains the variables.}

\item{x}{The binary variable to describe.}

\item{y}{The qualitative variable to perform the bivariate description. If unspecified, univariate description.}

\item{Prec}{Number of decimals for the percentages.}

\item{coefbin}{If TRUE displays the binomial confidence intervals.}

\item{chif_pval}{Number of decimals for the Pvalue if test is performed.}

\item{test}{String giving the name of the comparison test performed.}

\item{Langue}{"fr" for french and "eng" for english. For the display in the table.}

\item{nomcol}{Vector of strings to name each column of the output. Automatic display if unspecified.}

\item{nomcateg}{Sting giving the name of the class that you want to display in the table.}

\item{label}{The label you want to display for nomcateg.}

\item{simplif}{Not usefull, if TRUE will delete unused column 'Pvalue'.}
}
\description{
Description of binary variables
}
\details{
A function to describe binary variables with count and percentage (+ binomial confidence interval if supplied).

You can also use it to cross it with a categorial variable and perform a comparison test (chisq or fisher test).

The function has been updated to \code{TabBinaire}.
}
\examples{
tab_binaire(mtcars, am)
tab_binaire(mtcars, am, vs)
tab_binaire(mtcars, am, vs,
            nomcateg = "0", label = "Automatic transmission",
            test = "fisher")
}
\seealso{
\code{\link[=descr]{descr()}}, \code{\link[=TabBinaire]{TabBinaire()}}
}
