% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_quali.R
\name{tab_quali}
\alias{tab_quali}
\title{tab_quali}
\usage{
tab_quali(
  data,
  x,
  y = NULL,
  prec = 0,
  coefbin = FALSE,
  test = "none",
  chif_pval = 2,
  langue = "eng",
  nomcol = NULL,
  ordonnee = FALSE,
  nomvariable = NULL,
  simplif = TRUE
)
}
\arguments{
\item{data}{The dataset that contains the variables.}

\item{x}{The categorical variable to describe. Should be the name without "".}

\item{y}{The qualitative variable to perform the bivariate description. Should be the name without "". If unspecified, univariate description.}

\item{prec}{Number of decimals for the percentages.}

\item{coefbin}{If TRUE displays the binomial confidence intervals.}

\item{test}{String giving the name of the comparison test performed.}

\item{chif_pval}{Number of decimals for the Pvalue if test is performed.}

\item{langue}{"fr" for french and "eng" for english. For the display in the table.}

\item{nomcol}{Vector of strings to name each column of the output. Automatic display if unspecified.}

\item{ordonnee}{If FALSE, displays the classes of x in decreasing count order. If TRUE, displays in the original order.}

\item{nomvariable}{The Name of the variable you want to display. If not supplied, the name of the variable in data is used.}

\item{simplif}{Not usefull, if TRUE will delete unused column 'Pvalue'.}
}
\description{
Description of categorical variables.
}
\details{
A function to describe categorical variables with count and percentage (+ binomial confidence interval if supplied).

You can also use it to cross it with a categorial variable and perform a comparison test (chisq test).
}
\examples{
tab_quali(mtcars, cyl)
tab_quali(mtcars, cyl, am)
tab_quali(mtcars, cyl, am,
          nomvariable = "Number of cylinders",
          test = "chisq", chif_pval = 3)
}
