% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_quanti.R
\name{tab_quanti}
\alias{tab_quanti}
\title{tab_quanti}
\usage{
tab_quanti(
  data,
  x,
  y = NULL,
  prec = 0,
  chif_pval = 2,
  mode = "mediqrg",
  test = "none",
  langue = "eng",
  nomcol = NULL,
  nomvariable = NULL,
  simplif = TRUE
)
}
\arguments{
\item{data}{The dataset that contains the variables.}

\item{x}{The quantitative variable to describe.}

\item{y}{The qualitative variable to perform the bivariate description. If unspecified, univariate description.}

\item{prec}{Number of decimals for the statistics.}

\item{chif_pval}{Number of decimals for the Pvalue if needed.}

\item{mode}{String to indicate what statistics to display. If unspecified, by default give median, interquartile and range ("med" for
median, "moy" for mean, "sd" for standard deviation, "iq" for interquartile and "rg" fo range).}

\item{test}{String giving the test to compare the statistic between the 2 groups if needed. None if unspecified.}

\item{langue}{"fr" for french and "eng" for english. For the display in the table.}

\item{nomcol}{Vector of strings to name each column of the output. Automatic display if unspecified.}

\item{nomvariable}{String giving the name you want to display in the table. Automatic display if unspecified.}

\item{simplif}{Not usefull, if TRUE will delete unused column 'Pvalue'.}
}
\description{
Description of quantitative variables
}
\details{
A function to describe quantitative variables with mean, median, standard deviation, interquartile and range.

You can also use it to cross it with a categorial variable and perform a comparison test (student or wilcoxon test). For now no tests are performed with crossing with a variable that has more than 2 classes.
}
\examples{
tab_quanti(data = mtcars, x = mpg, prec = 1)
tab_quanti(data = mtcars, x = mpg, y = am, prec = 1)
tab_quanti(data = mtcars, x = mpg, y = am, prec = 1,
           nomvariable = "Miles per gallon",
           test = "stud", chif_pval = 3)
}
\seealso{
\code{\link{descr}}
}
