% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabQuali.R
\encoding{UTF-8}
\name{TabQuali}
\alias{TabQuali}
\title{TabQuali}
\usage{
TabQuali(
  .Data,
  x,
  y = NULL,
  Prec = 0,
  PMissing = NULL,
  NomVariable = NULL,
  ConfInter = c("none", "normal", "exact", "jeffreys"),
  ConfLevel = 0.95,
  Test = "none",
  ChifPval = 2,
  NomCol = NULL,
  Langue = "eng",
  Ordonnee = FALSE,
  Grapher = FALSE,
  Simplif = TRUE
)
}
\arguments{
\item{.Data}{The dataset that contains the variables.}

\item{x}{The categorical variable to describe. Should be the name without "".}

\item{y}{The qualitative variable to perform the bivariate description. Should be the name without "". If unspecified, univariate description.}

\item{Prec}{Number of decimals for the percentages.}

\item{PMissing}{Number of decimals of percentage from whole dataset. NULL (default value) if no percent desired.}

\item{NomVariable}{The Name of the variable you want to display. If not supplied, the name of the variable in data is used.}

\item{ConfInter}{Type of confidence interval (from normal, exact = Clopper-Pearson, and Jeffreys). None if no confidence interval is wanted.}

\item{ConfLevel}{Level of confidence for confidence intervals (by default 95\%).}

\item{Test}{String giving the name of the comparison test performed ("none", "chisq", "fisher").}

\item{ChifPval}{Number of decimals for the Pvalue if test is performed.}

\item{NomCol}{Vector of strings to name each column of the output. Automatic display if unspecified.}

\item{Langue}{"fr" for french and "eng" for english. For the display in the table.}

\item{Ordonnee}{If FALSE, displays the classes of x in decreasing count order. If TRUE, displays in the original order.}

\item{Grapher}{Boolean if you want to graph the distribution in univariate case.}

\item{Simplif}{Not usefull, if TRUE will delete unused column 'Pvalue'.}
}
\description{
Description of categorical variables
}
\details{
A function to describe categorical variables with count and percentage (+ binomial confidence interval if supplied).
The difference with \code{TabBinaire} is that all categories are displayed.

You can also use it to cross it with a categorial variable and perform a comparison test (chisq test, fisher test).
}
\examples{
TabQuali(mtcars, cyl)
TabQuali(mtcars, cyl, am)
TabQuali(mtcars, cyl, am,
         NomVariable = "Number of cylinders",
         Test = "chisq", ChifPval = 3)
}
\seealso{
\code{\link[=Description]{Description()}}
}
