% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabQuanti.R
\name{TabQuanti}
\alias{TabQuanti}
\title{TabQuanti}
\usage{
TabQuanti(
  .Data,
  x,
  y = NULL,
  Prec = 0,
  PMissing = NULL,
  NomVariable = NULL,
  Mode = "mediqrg",
  Test = "none",
  ChifPval = 2,
  NomCol = NULL,
  Langue = "eng",
  Grapher = FALSE,
  Simplif = TRUE
)
}
\arguments{
\item{.Data}{The dataset that contains the variables.}

\item{x}{The quantitative variable to describe.}

\item{y}{The qualitative variable to perform the bivariate description. If unspecified, univariate description.}

\item{Prec}{Number of decimals for the statistics.}

\item{PMissing}{Number of decimals of percentage from whole dataset. NULL (default value) if no percent desired.}

\item{NomVariable}{String giving the name you want to display in the table. Automatic display if unspecified.}

\item{Mode}{String to indicate what statistics to display. If unspecified, by default give median, interquartile and range ("med" for
median, "moy" for mean, "sd" for standard deviation, "iq" for interquartile and "rg" fo range).}

\item{Test}{String giving the test to compare the statistic between the 2 groups if needed ("none", "ztest", "student", "wilcoxon").}

\item{ChifPval}{Number of decimals for the Pvalue if needed.}

\item{NomCol}{Vector of strings to name each column of the output. Automatic display if unspecified.}

\item{Langue}{"fr" for french and "eng" for english. For the display in the table.}

\item{Grapher}{Boolean if you want to graph the distribution in univariate case.}

\item{Simplif}{Not usefull, if TRUE will delete unused column 'Pvalue'.}
}
\description{
Description of quantitative variables
}
\details{
A function to describe quantitative variables with mean, median, standard deviation, interquartile and range.

You can also use it to cross it with a categorial variable and perform a comparison test (student or wilcoxon test). For now no tests are performed with crossing with a variable that has more than 2 classes.
}
\examples{
TabQuanti(.Data = mtcars, x = mpg, Prec = 1)
TabQuanti(.Data = mtcars, x = mpg, y = am, Prec = 1)
TabQuanti(.Data = mtcars, x = mpg, y = am, Prec = 1,
          NomVariable = "Miles per gallon",
          Test = "stud", ChifPval = 3)
}
\seealso{
\code{\link[=Description]{Description()}}
}
