% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{ThemeDescription}
\alias{ThemeDescription}
\title{Flextable theme for description table}
\usage{
ThemeDescription(tab, BaseSize = 10, ColNum = NULL, Separations = NULL)
}
\arguments{
\item{tab}{Flextable object.}

\item{BaseSize}{Size of police.}

\item{ColNum}{Index of numerical columns.}

\item{Separations}{Index of columns we want separator.}
}
\value{
A flextable object (first column and headers can be written in markdown like synthax).
}
\description{
Flextable theme for description table
}
