% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{ThemePerso}
\alias{ThemePerso}
\title{Flextable personal theme}
\usage{
ThemePerso(tab, BaseSize = 10)
}
\arguments{
\item{tab}{Flextable objet.}

\item{BaseSize}{Size of police.}
}
\value{
A flextable object.
}
\description{
Flextable personal theme
}
