% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verif_mode}
\alias{verif_mode}
\title{Check the statistical presentation}
\usage{
verif_mode(x)
}
\arguments{
\item{x}{String to indicate the presentation}
}
\description{
Check the statistical presentation
}
