% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verif_test}
\alias{verif_test}
\title{Check the statistical test}
\usage{
verif_test(x, type_var)
}
\arguments{
\item{x}{String naming the test.}

\item{type_var}{Type of variable.}
}
\description{
Check the statistical test
}
