% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{VerifArgs}
\alias{VerifArgs}
\title{Check the validity of supplied arguments for the functions}
\usage{
VerifArgs(...)
}
\arguments{
\item{...}{The tested argument with helping arguments if needed. The first argument will guide the function}
}
\value{
The value of the argument
}
\description{
Check the validity of supplied arguments for the functions
}
