% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.R
\name{VerifTest}
\alias{VerifTest}
\title{Check the statistical test}
\usage{
VerifTest(Test, TypeVar, NClasses, Variable, y, x)
}
\arguments{
\item{x}{String naming the test.}

\item{type_var}{Type of variable.}
}
\description{
Check the statistical test
}
