% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_data.R
\name{visu_data}
\alias{visu_data}
\title{visu_data}
\usage{
visu_data(data, datavar, langue = "fr")
}
\arguments{
\item{data}{The data.frame of the dataset you want to explore.}

\item{datavar}{The data.frame of the datavar containing metadata on the dataset.}

\item{langue}{For now, "fr" for french or "eng" for english.}
}
\description{
Rapid visualisation of the data with the informations given by the datavar
}
\details{
Open a simple shinyapp to graphically explore the variables. If a plot is satisfactory enough, you can copy paste the code to reproduce it. You'll need to load library \code{ggplot2}.
}
\examples{
visu_data(data = mtcars, datavar = datavarr, langue = "eng")
}
