% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WeightedQuantile}
\alias{WeightedQuantile}
\title{Weighted quantiles}
\usage{
WeightedQuantile(x, w = rep(1, length(x)), q = 0.5)
}
\description{
Weighted quantiles
}
\references{
https://en.wikipedia.org/wiki/Weighted_median
}
