% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WeightedVar}
\alias{WeightedVar}
\title{Weighted variance}
\usage{
WeightedVar(x, w, na.rm = TRUE, Corr = 1)
}
\description{
Weighted variance
}
\references{
https://www.gnu.org/software/gsl/doc/html/statistics.html#weighted-samples
https://www.analyticalgroup.com/download/weighted_mean.pdf
}
