% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{ZTest}
\alias{ZTest}
\title{Z-test}
\usage{
ZTest(
  VarQuanti,
  VarY = NULL,
  Paired = FALSE,
  Alternative = c("two.sided", "less", "greater"),
  Mu = 0,
  ConfLevel = 0.95
)
}
\arguments{
\item{VarQuanti}{Numeric vector of data.}

\item{VarY}{Either numeric vector of paired data with VarQuanti or binary variable indicating the groups to compare.}

\item{Paired}{TRUE for paired Z-test.}

\item{Alternative}{Character string for alternative hypothesis ("two.sided", "less", "greater").}

\item{Mu}{A numeric indicating the value to which we want to compare the mean in one-sample test or difference of means in two-sample test.}

\item{ConfLevel}{Confidence level of the interval of the difference.}
}
\value{
A list of class "htest" containing the following components :
\describe{
\item{statistic}{The value of Z-statistic.}
\item{p.value}{PValue of the test.}
\item{conf.int}{Confidence interval of the difference of means.}
\item{estimate}{Estimated mean or difference in means.}
\item{null.value}{The value to which we want to compare the mean in one-sample test or difference of means in two-sample test.}
\item{stderr}{Standard error of the difference in means.}
\item{alternative}{Alternative hypothesis.}
\item{method}{Type of Z-test.}
\item{data.name}{Name(s) of the data.}
}
}
\description{
Perform one- and two-samples Z-tests.
}
\examples{
ZTest(mtcars$mpg, mtcars$vs)
}
