% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datavar.R
\name{create_datavar}
\alias{create_datavar}
\title{create_datavar}
\usage{
create_datavar(
  data,
  colonnes = colnames(data),
  default_datavar = options_datavar()
)
}
\arguments{
\item{data}{The dataset for which we want a datavar.}

\item{colonnes}{The columns of the dataset that you want to describe. By default, all the columns of the dataset are to be described.}

\item{default_datavar}{The default options for the creation. Made by the function \code{options_datavar}}
}
\value{
A data.frame which is the datavar.
}
\description{
Creation of a default datavar
}
\seealso{
\code{\link{descr}}
}
