% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabBinaire.R
\name{TabBinaire}
\alias{TabBinaire}
\title{TabBinaire}
\usage{
TabBinaire(
  .Data,
  x,
  y = NULL,
  Prec = 0,
  PMissing = NULL,
  NomCateg = NULL,
  NomLabel = NULL,
  ConfInter = c("none", "normal", "exact", "jeffreys"),
  ConfLevel = 0.95,
  Test = "none",
  ChifPval = 2,
  NomCol = NULL,
  Langue = "eng",
  Grapher = FALSE,
  Simplif = FALSE
)
}
\arguments{
\item{.Data}{The dataset that contains the variables.}

\item{x}{The binary variable to describe.}

\item{y}{The qualitative variable to perform the bivariate description. If unspecified, univariate description.}

\item{Prec}{Number of decimals for the percentages.}

\item{PMissing}{Number of decimals of percentage from whole dataset. NULL (default value) if no percent desired.}

\item{NomCateg}{The value of the category which you want to display in the table.}

\item{NomLabel}{String giving the name of the class that you want to display in the table.}

\item{ConfInter}{Type of confidence interval (from normal, exact = Clopper-Pearson, and Jeffreys). None if no confidence interval is wanted.}

\item{ConfLevel}{Level of confidence for confidence intervals (by default 95\%).}

\item{Test}{String giving the name of the comparison test performed ("none", "chisq", "fisher").}

\item{ChifPval}{Number of decimal for PValue.}

\item{NomCol}{Vector of strings to name each column of the output. Automatic display if unspecified.}

\item{Langue}{"fr" for french and "eng" for english. For the display in the table.}

\item{Grapher}{Boolean if you want to graph the distribution in univariate case.}

\item{Simplif}{Boolean. If TRUE (default value) the table will be simplified to remove number of missing data if there aren't any.}
}
\description{
Description of binary variables
}
\details{
A function to describe binary variables with count and percentage (+ binomial confidence interval if supplied).
The difference with \code{TabQuali} is that \code{TabBinaire} only displays results for the wanted category supplied.

You can also use it to cross it with a categorial variable and perform a comparison test (chisq or fisher test).
}
\examples{
TabBinaire(mtcars, am)
TabBinaire(mtcars, am, vs)
TabBinaire(mtcars, am, vs,
           NomCateg = "0", NomLabel = "Automatic transmission",
           Test = "fisher")
}
\seealso{
\code{\link[=Description]{Description()}}
}
