% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description.R
\name{VarDescr}
\alias{VarDescr}
\title{VarDescr}
\usage{
VarDescr(
  .Data,
  LigneDatavar,
  y,
  NomCol,
  PMissing = NULL,
  Poids = NULL,
  SMD = FALSE,
  Simplif = TRUE,
  Langue,
  Grapher = FALSE
)
}
\arguments{
\item{.Data}{Data.frame of the data.}

\item{LigneDatavar}{The line in datavar of the variable to describe.}

\item{y}{The variable used to cross.}

\item{NomCol}{Vector of the names of the columns.}

\item{PMissing}{Number of decimals of percentage from whole dataset. NULL (default value) if no percent desired.}

\item{Poids}{Name of the column of .Data in which the weights are stored. Let NULL for unweighted analysis.}

\item{SMD}{Boolean to indicate if you want standardized mean differences. Of note, for weighted analysis, only SMD are available and no test.}

\item{Simplif}{Boolean. If TRUE (default value) the table will be simplified to remove number of missing data if there aren't any.}

\item{Langue}{Language.}

\item{Grapher}{Boolean if you want to graph the distribution in univariate case.}
}
\description{
VarDescr
}
